pyproject.toml
[project]
name = "vibe-coding-template"
version = "0.1.0"
description = "A template for Vibe Coding data science projects."
authors = [{ name = "Connor Kitchings", email = "connor.kitchings@gmail.com" }]
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "duckdb>=0.10.0",
    "fredapi>=0.5.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "tenacity>=9.1.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.0",
]

# Optional: Add these as needed for your project
data-science = [
    "jupyter>=1.0",
    "nbstripout>=0.7.1",
    "nbformat>=5.10.4",
]

mlops = [
    "mlflow>=2.0",
    "prefect>=2.19.0",
    "openlineage-python>=1.15.0",
]

security = [
    "bandit>=1.7.9",
    "safety>=2.3.5",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "--cov=vibe_coding",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-fail-under=55",
    "-v",
]

[tool.coverage.run]
source = ["src/vibe_coding"]
branch = true
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/conftest.py",
]

[tool.coverage.report]
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pytest-cov>=7.0.0",
    "ruff>=0.14.3",
]
